Created a resource helper static class to store info for get and set resource table values
